requirements.txt
setuptools~=75.1.0
requests~=2.32.3
beautifulsoup4~=4.12.3
paramiko~=2.8.1
pyqt5
scapy~=2.4.3
dnspython~=2.4.2
boto3~=1.34.154
botocore~=1.34.154
instaloader~=4.13.1